fix: split manual scenario integration tests for ui-lint file length
