+++ docs/internal/processes/api-relevant-files.md
@@ -98,7 +98,7 @@ Use this map when changing the public REST contract.
 - Shared app-shell dashboard tests that render `<App />` through `ui/src/testing/app-shell-test-utils.tsx` also depend on that preflight route now; keep the default fetch harness answering `GET /factory-sessions/{session_id}/sync-preflight` with a valid typed response so replay/current-selection integration tests exercise the real dashboard shell instead of stalling in preflight recovery because the mock API forgot the new contract.
 - App-shell integration tests that need custom `fetch` behavior should layer it through `chainRenderAppFetchMock(...)` in `ui/src/testing/app-shell-test-utils.tsx` instead of replacing `fetchMock` wholesale; overriding the mock directly can accidentally drop the shared `sync-preflight` route and boot the shell into recovery instead of the state the test intends to exercise.
 - Browser-backed dashboard integration fixtures in `ui/integration/browser-test-harness.mjs` also need the sync-preflight route now; keep `startFactoryApiServer(...)` serving `GET /factory-sessions/{session_id}/sync-preflight` with the typed identity fields and production status semantics the dashboard requires. In particular, typed outcomes such as `reasonCode=session_not_found` still return HTTP `200`; do not turn them into non-OK transport failures or the browser lane will prove the wrong recovery path and can burn the timeout before failing.
-- Dashboard checkpoint identity recovery UI proofs live in `ui/src/features/dashboard/hooks/useDashboardSnapshot.test.tsx` (matching identity cursor reuse, backend/stream/remap mismatches, identity-missing checkpoint deletion, sanitized mismatch diagnostics), `ui/src/features/dashboard/components/dashboard-screen.test.tsx` and `dashboard-screen.recovery-retry.test.tsx` (distinct recoverable replay failure vs generic offline chrome, keyboard-reachable retry), `ui/integration/dashboard-session-recovery.integration.test.mjs` (browser-visible checkpoint cursor reuse, stale-identity reconnect without cursor, recovery panel at desktop and narrow viewports), and `ui/integration/dashboard-session-recovery-manual-scenarios.integration.test.mjs` (restart, clean-restart, remap, backend-scope switch, and multi-tab checkpoint invalidation with captured `EventSource` URLs). Manual scenario evidence with commands and sanitized identity outcomes lives in `docs/internal/development/session-persistence-restart-switching-evidence.md`. Seed IndexedDB checkpoints only after the preview origin is loaded; clear persisted checkpoints between browser scenarios when later reloads must not reuse an earlier matching identity envelope.
+- Dashboard checkpoint identity recovery UI proofs live in `ui/src/features/dashboard/hooks/useDashboardSnapshot.test.tsx` (matching identity cursor reuse, backend/stream/remap mismatches, identity-missing checkpoint deletion, sanitized mismatch diagnostics), `ui/src/features/dashboard/components/dashboard-screen.test.tsx` and `dashboard-screen.recovery-retry.test.tsx` (distinct recoverable replay failure vs generic offline chrome, keyboard-reachable retry), `ui/integration/dashboard-session-recovery.integration.test.mjs` (browser-visible checkpoint cursor reuse, stale-identity reconnect without cursor, recovery panel at desktop and narrow viewports), and `ui/integration/dashboard-session-recovery-manual-scenarios.integration.test.mjs` plus `ui/integration/dashboard-session-recovery-manual-scenarios-switching.integration.test.mjs` (restart, clean-restart, remap, backend-scope switch, provider/account-scope switch, and multi-tab checkpoint invalidation with captured `EventSource` URLs). Manual scenario evidence with commands and sanitized identity outcomes lives in `docs/internal/development/session-persistence-restart-switching-evidence.md`. Seed IndexedDB checkpoints only after the preview origin is loaded; clear persisted checkpoints between browser scenarios when later reloads must not reuse an earlier matching identity envelope.
 - Non-recoverable typed preflight outcomes for dashboard restore belong in `ui/src/features/dashboard/lib/dashboard-session-sync-preflight.ts` and flow through `useDashboardSnapshot` into `ui/src/features/dashboard/components/dashboard-screen.tsx`; when `reasonCode=session_not_found` (or another typed non-recoverable outcome), suppress the generic offline shell, keep the dashboard header visible, and render a clean-replay recovery action instead of opening the event stream.
 - If the dashboard advertises a clean-replay recovery action for a failed sync preflight request, `useDashboardCheckpointPreflight.ts` must clear the persisted timeline checkpoint before the retry path runs again; otherwise the next preflight will silently resend the stale reconnect cursor even though the UI claims cached replay data was cleared.
 - `pkg/api/contracttests/` contains OpenAPI authoring, bundled-contract, and generated-type smoke tests for public API changes. Residual active-surface import gate verification uses scoped `rg` commands documented in `docs/internal/development/plans/dynamic-workflows/dynamic-workflows-plan.md` plus behavioral preview/alias tests in contract and server packages.
